requirements.txt
langchain==0.1.10
langchain-openai==0.0.8
langchain-google-genai==0.0.9
langgraph==0.0.25
pylint==3.0.3
pytest==7.4.4
python-dotenv==1.0.1
pandas==2.2.0
colorama==0.4.6
google-generativeai>=0.3.1,<0.4.0